+++ app/helpers/ImageUpload_Helper.php
@@ -46,11 +46,15 @@ function updateImage($old, $img, $img_name, $location) {
 }
 
 
-function deleteImage($img){
-    if(unlink($img)){
-        return true;
+function deleteImage($img) {
+    if (basename($img) === 'profile.png') {
+        die();
+        return false; // Don't delete if the filename is "profile.png"
     }
-    else{
+
+    if (unlink($img)) {
+        return true;
+    } else {
         return false;
     }
 }
